# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Han Sports v2 API
  version: 1.0.0
  description: Starter OpenAPI contract for the Han Sports v2 Spring Boot REST API.
servers:
  - url: http://localhost:8080
    description: Local backend
paths:
  /api/v1/auth/register:
    post:
      summary: Register a user account
      tags: [Auth]
      responses:
        "201":
          description: User registered
        "400":
          description: Validation or duplicate email error
  /api/v1/auth/login:
    post:
      summary: Login and receive an access token plus refresh cookie
      tags: [Auth]
      responses:
        "200":
          description: Login success
        "400":
          description: Invalid credentials
        "429":
          description: Rate limit exceeded
  /api/v1/auth/refresh:
    get:
      summary: Rotate refresh token and issue a new access token
      tags: [Auth]
      responses:
        "200":
          description: Token refreshed
        "401":
          description: Missing or invalid refresh cookie
  /api/v1/auth/logout:
    post:
      summary: Revoke refresh token and clear cookie
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout success
  /api/v1/auth/change-password:
    post:
      summary: Change the current user's password
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Password changed and refresh token revoked
        "400":
          description: Invalid current password or invalid new password
  /api/v1/products:
    get:
      summary: List products with pagination and optional filter
      tags: [Products]
      responses:
        "200":
          description: Product page
    post:
      summary: Create a product
      tags: [Products]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Product created
        "403":
          description: Admin role required
    put:
      summary: Update a product
      tags: [Products]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Product updated
        "403":
          description: Admin role required
  /api/v1/products/{id}:
    get:
      summary: Get product detail
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
        "400":
          description: Product not found
    delete:
      summary: Delete a product
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
        "403":
          description: Admin role required
  /api/v1/files:
    get:
      summary: Download an uploaded file
      tags: [Files]
      responses:
        "200":
          description: File stream
    post:
      summary: Upload one or more image files
      tags: [Files]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: File uploaded
        "400":
          description: Invalid image file
        "403":
          description: Admin role required
  /api/v1/carts:
    get:
      summary: Get current user's cart
      tags: [Cart]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart detail list
    post:
      summary: Add a product to cart
      tags: [Cart]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart updated
  /api/v1/carts/{id}:
    put:
      summary: Update cart detail quantity
      tags: [Cart]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quantity updated
        "400":
          description: Quantity invalid or exceeds stock
    delete:
      summary: Remove cart detail
      tags: [Cart]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
  /api/v1/orders:
    get:
      summary: Admin list all orders
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Order page
        "403":
          description: Admin role required
    post:
      summary: Create an order from selected cart details
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Order created
        "400":
          description: Invalid cart or insufficient stock
    put:
      summary: Admin update order status
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Order updated
        "403":
          description: Admin role required
  /api/v1/orders/{id}/send-email:
    post:
      summary: Admin send order confirmation email
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email sent
        "403":
          description: Admin role required
  /api/v1/settings:
    get:
      summary: List public application settings
      tags: [Settings]
      responses:
        "200":
          description: Settings list
    put:
      summary: Admin bulk update settings
      tags: [Settings]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings updated
        "400":
          description: Invalid setting payload
        "403":
          description: Admin role required
  /actuator/health:
    get:
      summary: Application health check
      tags: [Operations]
      responses:
        "200":
          description: Health status
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT